requirements.txt
pygame
pyinstaller
pycryptodome
tkinter
